move around vap and rename
