+++ protocol-designer/src/steplist/selectors.js
@@ -6,7 +6,12 @@ import mapValues from 'lodash/mapValues'
 import max from 'lodash/max'
 
 import {selectors as labwareIngredSelectors} from '../labware-ingred/reducers'
-import {getFormWarnings, getFormErrors} from './formLevel'
+import {
+  getFormWarnings,
+  getFormErrors,
+  generateNewForm,
+  stepFormToArgs
+} from './formLevel'
 import type {FormError, FormWarning} from './formLevel'
 import {hydrateField} from './fieldLevel'
 import {initialSelectedItemState} from './reducers'
@@ -26,12 +31,7 @@ import type {
   StepIdType
 } from '../form-types'
 
-import {
-  type ValidFormAndErrors,
-  generateNewForm,
-  validateAndProcessForm,
-  formHasErrors
-} from './formProcessing'
+import { type ValidFormAndErrors } from './formLevel/stepFormToArgs'
 
 // TODO Ian 2018-01-19 Rethink the hard-coded 'steplist' key in Redux root
 const rootSelector = (state: BaseState): RootState => state.steplist
@@ -140,7 +140,7 @@ const validatedForms: Selector<{[StepIdType]: ValidFormAndErrors}> = createSelec
   (_steps, _savedStepForms, _orderedSteps) => {
     return reduce(_orderedSteps, (acc, stepId) => {
       const nextStepData = (_steps[stepId] && _savedStepForms[stepId])
-        ? validateAndProcessForm(_savedStepForms[stepId])
+        ? stepFormToArgs(_savedStepForms[stepId])
         // NOTE: usually, stepFormData is undefined here b/c there's no saved step form for it:
         : {
           errors: {'form': ['no saved form for step ' + stepId]},
@@ -246,7 +246,7 @@ const nextStepId: Selector<number> = createSelector( // generates the next step 
 // TODO: remove this when we add in form level validation
 const currentFormErrors: Selector<null | {[errorName: string]: string}> = (state: BaseState) => {
   const form = formData(state)
-  return form && validateAndProcessForm(form).errors // TODO refactor selectors
+  return form && stepFormToArgs(form).errors // TODO refactor selectors
 }
 
 const formLevelWarnings: Selector<Array<FormWarning>> = (state) => {
@@ -315,15 +315,14 @@ const getSelectedStep = createSelector(
   }
 )
 
+// TODO: BC: 2018-08-21 remove this, always allow save
 export const currentFormCanBeSaved: Selector<boolean | null> = createSelector(
   formData,
   getSelectedStepId,
   allSteps,
   (formData, selectedStepId, allSteps) =>
     ((typeof selectedStepId === 'number') && allSteps[selectedStepId] && formData)
-      ? !formHasErrors(
-        validateAndProcessForm(formData)
-      )
+      ? Object.values(stepFormToArgs(formData).errors).length === 0
       : null
 )
